- cleanup visibility
