+++ src/Controls/src/Core/SwipeItem.cs
@@ -23,15 +23,5 @@ namespace Microsoft.Maui.Controls
 		}
 
 		public event EventHandler<EventArgs> Invoked;
-
-		[EditorBrowsable(EditorBrowsableState.Never)]
-		public void OnInvoked()
-		{
-			if (Command != null && Command.CanExecute(CommandParameter))
-				Command.Execute(CommandParameter);
-
-			OnClicked();
-			Invoked?.Invoke(this, EventArgs.Empty);
-		}
 	}
 }
