+++ src/Core/src/Core/ISwipeItem.cs
@@ -1,0 +1,11 @@
+﻿namespace Microsoft.Maui
+{
+	public interface ISwipeItem : IElement
+	{
+		void OnInvoked();
+
+		string AutomationId { get; }
+	}
+
+
+}
