+++ src/Core/src/Core/ISwipeItemMenuItem.cs
@@ -1,0 +1,8 @@
+﻿namespace Microsoft.Maui
+{
+	public interface ISwipeItemMenuItem : IMenuElement, ISwipeItem
+	{
+	}
+
+
+}
