+++ src/Core/src/Core/ISwipeItemView.cs
@@ -1,0 +1,6 @@
+﻿namespace Microsoft.Maui
+{
+	public interface ISwipeItemView : IContentView, ISwipeItem
+	{
+	}
+}
